+++ src/json.ts
@@ -4,9 +4,8 @@ import {Codec} from "./codec";
 export type Json = boolean | number | string | any[] | {[k: string]: any};
 
 export type JsonPrimitive = boolean | number | string;
-export type JsonObject = {[prop: string]: Json | null};
+export type JsonObject<T> = {[K in keyof T]: Json | null };
 export type JsonArray  = Array<Json>;
-
 
 export const bool: Codec<boolean,boolean> = nullSafeIdentity;
 export const num: Codec<number,number> = nullSafeIdentity;